Refactor Navbar, Fix content typos, Reoder Sections
Split Navbar code, And fix the content headers and fix 2-line contents
