+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "workbench.colorTheme": "Material Theme"
+  "workbench.colorTheme": "Material Theme Darker High Contrast"
 }
