+++ src/content/smartSaverBannerContent.js
@@ -1,5 +1,5 @@
 const smartSaverBannerContent = {
-  title: 'Education is the key to brighter future.',
+  title: 'Education is the key to a brighter future.',
   subtitle: 'Sending your child to higher education can be expensive.',
   content:
     'There are two things you can do now to help your child to reach their potential (without costing you a dime!)',
